- Added upsert for new users - redirect to home and not to the artist page by default - minor styles fixes - error handling in services
